rip npm_and_my_sanity
