+++ examples/ccxt_and_gravatar/Cargo.toml
@@ -1,14 +1,13 @@
 [package]
-name = "npm_and_rest"
+name = "ccxt_and_gravatar"
 version = "0.1.0"
 authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-anyhow = "1"
+anyhow = "1.0"
 js-sys = "0.3"
-serde = "1"
-serde_derive = "1"
+serde = { version = "1.0", features = ["derive"] }
 wasm-bindgen = "0.2"
 web-sys = { version = "0.3", features = ["console"] }
 yew = { path = "../../yew" }
