+++ examples/ccxt_and_gravatar/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew • CCXT and Gravatar</title>
+
+    <script src="https://unpkg.com/ccxt"></script>
+  </head>
+
+  <body></body>
+</html>
