--- examples/npm_and_rest/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    yew::start_app::<npm_and_rest::Model>();
-}
