--- examples/npm_and_rest/static/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <title>Yew • npm and REST</title>
-        <script type="text/javascript" src="https://unpkg.com/ccxt"></script>
-    </head>
-    <body>
-        <script type="module">
-            import init from "./wasm.js"
-            init()
-        </script>
-    </body>
-</html>
